Create CodeGenBase class
